requirements.txt
langchain
langchain-ollama
langchain-chroma
python-dotenv
requests